+++ Changelog.md
@@ -8,6 +8,7 @@ Compiler Features:
 
 Bugfixes:
  * Code Generator: Fix internal compiler error when calling functions bound to calldata structs and arrays.
+ * Type Checker: Fix internal error and prevent static calls to unimplemented modifiers.
 
 
 ### 0.8.6 (2021-06-22)